requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.0.6
cprint==1.2.2
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1652361911092/work
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
iopath==0.1.9
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kiwisolver==1.4.4
kornia==0.6.8
lightning-utilities==0.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020388495/work
numpy-quaternion==2022.4.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1666867978941/work
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch-lightning==1.8.3.post1
pytorch3d==0.7.1
pytz==2022.6
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
pyzmq==24.0.1
Quaternion==3.5.2.post4
requests==2.28.1
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1665138452165/work
tensorboardX==2.5.1
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1667338503774/work
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.9.1
torchmetrics==0.10.3
torchvision==0.10.1
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets==5.6.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.13
wandb==0.13.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
yarl==1.8.1
zipp==3.11.0
